Bug 1176650 - Refactored fetching client IPs.
This does a bunch of things:

- Introduce a middleware to use the guarenteed client IP from the load balancer via the X-Forwarded-For header.

- Combined the two existing tools that defer a unique identifier from the request.

- Cleaned up a few tests.
